mm: restore full accuracy in COW page reuse
This reverts commit 09854ba94c6aad7886996bfbee2530b3d8a7f4f4.

After a GUP pin is taken, changing the physical address of the page
mapped in the pagetable causes the holder of the GUP pin to lose
synchrony with the CPU. It's a feature that a GUP pinned page acts as
an anchor for the physical address mapped in the pagetable and
prevents it from being copied by a COW fault.

The page_count check in do_wp_page achieves the exact opposite: it
guarantees a GUP pinned page is copied during a wrprotect fault, even
if it is exclusive and not shared with other processes.

Checking page_count instead of page_mapcount in do_wp_page makes it
impossible to wrprotect any page under a readonly long term GUP
pin. Otherwise during the wrprotect fault the physical address in the
pagetable would risk to change causing all GUP long term pins to lose
synchrony with any update from the CPU, right after the wrprotect
fault has been resolved.

The same applies to the FOLL_PIN because from the MM standpoint there
is no difference if the pin is taken with FOLL_GET or FOLL_PIN.

However, as shown in the below example, even short term GUP pins start
exhibiting unexpected results with the page_count check. The example
below uses O_DIRECT, on a process under clear_refs. There's no long
term GUP pin here. All GUP pins are transient. fork is never called
and even when clear_refs is cleared by an external program, fork()
would not be involved.

thread0				thread1			other process
							(or thread 3)
=============			=============		===========
read syscall
O_DIRECT
read DMA to vaddr+0
len = 512 bytes
GUP(FOLL_WRITE)
DMA writing to RAM started
							clear_refs
							pte_wrprotect
				write vaddrA+512
				page_count == 2
				wp_copy_page
read syscall returns
read lost

Notwithstanding the fact that failing O_DIRECT at sub-PAGE_SIZE
granularity is an ABI break, recvmsg kernel TLS and plenty of other
GUP FOLL_WRITE iov_iter_get_pages users would write to the memory with
sub-PAGE_SIZE granularity too. Those recvmsgs are silently lost too as
well as the above O_DIRECT gets lost.

That COW must not happen too much is documented in commit
6d0a07edd17cfc12fdc1f36de8072fa17cc3666f:

==
This will provide fully accuracy to the mapcount calculation in the
write protect faults, so page pinning will not get broken by false
positive copy-on-writes.
==

And in the current comment above the THP mapcount:

==
 * [..] we only use
 * page_trans_huge_mapcount() in the copy-on-write faults where we
 * need full accuracy to avoid breaking page pinning, [..]
==

This revert causes no theoretical security regression because THP is
not yet using page_count in do_huge_pmd_wp_page.

The alternative of this patch would be to replace the mapcount with
the page_count in do_huge_pmd_wp_page too in order to really close the
vmsplice long term GUP attack from child to parent that way.

However, because a single transient GUP pin on a tail page would
elevate the page_count for all other tail pages (unlike the mapcount
which is single page granular), if the COW page reuse inaccuracy would
be applied to THP too, it could cross different vmas and the effect
could happen at a distance in vmas of different processes. A single
GUP pin taken on a subpage mapped in a different process could trigger
511 false positive COWs copies in the local process in the worst case,
after a fork(). Which would be yet an additional inefficiency on top
of the previous cons.

Another non functional, but purely performance related regression
caused by the spurious COW copies of exclusive anonymous memory, is
that we cannot assume anymore that deferred TLB flushes after
wrprotection are noops with respect to the COW fault. The stale
writable TLBs can be seen as the equivalent of a driver holding a GUP
pin on exclusive (non shared) anonymous memory. If spurious COW copies
can happen, the physical address of the page will change after the
copy. So before the COW fault starts copying the memory to the new
physical location, to avoid losing writes on the page, the stale
writable TLBs have to be invalidated. This is non a functional
regression because it is always possible to "flush the TLB" at any
time (or alternatively to wait for the deferred TLB flushes to be
executed). However, what's not possible is to "flush the GUP pin" (nor
to wait for it to be released).

Link: https://lkml.kernel.org/r/20210107200402.31095-1-aarcange@redhat.com
Fixes: 09854ba94c6a ("mm: do_wp_page() simplification")
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>