mm: COW: skip the page lock in the COW copy path
The only guarantee required to be allowed to copy is that mapcount is
not equal 1. That is enough not to risk overestimating mapcount.

This should retain the performance improvement reported below for the
COW copy path, but without introducing any regression to the page
reuse accuracy.

https://lore.kernel.org/lkml/20200914024321.GG26874@shao2-debian/

In addition this optimizes the THP COW with the same logic.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>