mm: thp: make the THP total_mapcount atomic against rmap add/remove
Use the page_mapcount_lock in the rmap add/remove to keep the
DoubleMap atomic with the compound_mapcount and the mapcount.

FOLL_NOSHARE requires to read the THP mapcount not just in gup-fast
without page lock, but also from gup-fast-only which may run in irq
context and needs an irq safe atomic reading of the THP mapcount.

Reported-by: Hugh Dickins <hughd@google.com>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>