mm: gup: COR: copy-on-read fault
COR is a new type of page fault that generates readonly anonymous
memory and its only purpose is to allow the unsharing of
write-protected COW shared pages, while still maintaining their write
protection. It is only meant to be used within GUP and it can be
invoked through the FAULT_FLAG_UNSHARE flag.

Reviewed-by: Peter Xu <peterx@redhat.com>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>