KSM: gup: FOLL_UNSHARE: copy-on-read fault
Combined with the enforced invariant that an already GUP pinned
PageAnon cannot be merged as PageKsm, this enforces that no GUP pin
can be present on a PageKsm.

In turn the current page_count check in do_wp_page if applied only to
PageKsm, won't risk breaking any GUP pin anymore.

Lack of this GUP fix for PageKsm, could have been worked around by
FOLL_LONGTERM callers by combining it with FOLL_FORCE|FOLL_WRITE even
for readonly GUPs. That's not needed anymore also for PageKsm.

NOTE: it would be theoretically possible to GUP pin a Ksm page with
mapcount == 1, but other invariant would break within KSM because
further merges would need to be disallowed and it shall remain
exclusive, so it's just simpler to implement the un-share of a PageKsm
with page_mapcount in zerocopy with reuse_page_ksm than to try to
handle that case.