mm: thp: optimize compound_nr
Minor cleanup and minor optimization.

There's no point to return "unsigned long" in compound_nr, when it is
an int in the page structured and in places it was even assigned to a
int variable. In addition the value should be allowed to remain in a
register and not be forced to be re-read from memory at every
iteration of the loop even if all functions called aren't in scope or
inline.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>