mm: mprotect: avoid spurious COW faults for exclusive anon pages in cow mapping
For cow mappings (anon vmas and MAP_PRIVATE of filebacked vmas), if
the page is anon, the mapcount shows the anon page is exclusive, if
the vmas has VM_WRITE set we can mark the pte writable.

In the THP path the page_mapcount is too slow, a boolean retval is
allowed to return "non exclusive" without checking every subpage so
page_trans_huge_anon_shared() is used as an optimization instead of
page_mapcount() != 1, but otherwise they remain identical.

The THP mapcount is already atomic, it uses the readonly seqlock
internally, so there's no need to take any page lock in the mprotect
fast path for the THP path to skip the spurious COWs on THPs.

Original idea and patch posted here:

https://lkml.kernel.org/r/20210429214801.2583336-1-pcc@google.com

Fix for uffd-wp from Peter Xu:

https://lkml.kernel.org/r/YKgCsmMVp3wNLLBB@t490s

Suggested-by: Peter Collingbourne <pcc@google.com>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>