randprotect: KSM: mimic the timing of the COW break for VM_MERGEABLE vmas
eBPF tracing can be used on a testcase to measure the COW fault
latency for matched and unmatched pages.

==== ksm-writeprotect.bp ====
 #!/usr/bin/env bpftrace

 kprobe:handle_mm_fault /comm == "ksm-COW"/
 {
 	@start[tid] = nsecs;
 }
 kprobe:do_wp_page /comm == "ksm-COW" && @start[tid]/
 {
 	@do_wp_page[tid] = 1;
 }
 kretprobe:handle_mm_fault /comm == "ksm-COW" && @start[tid]/
 {
 	if (@do_wp_page[tid]) {
 		@latency[tid] = hist(nsecs - @start[tid]);
 	}
 	delete(@start[tid]);
 	delete(@do_wp_page[tid]);
 }
==== ksm-COW-lat.c ====
 #include <string.h>
 #include <sys/mman.h>
 #include <stdlib.h>
 #include <unistd.h>
 #include <stdio.h>

 #define PAGE_SIZE 4096
 #define SIZE (16*PAGE_SIZE)

 int main(int argc, char *argv[])
 {
 	char *p;
 	if (posix_memalign((void **)&p, PAGE_SIZE, SIZE))
 		perror("malloc"), exit(1);
 	if (madvise(p, SIZE, MADV_MERGEABLE))
 		perror("madvise"), exit(1);
 	if (argc > 1) {
 		for (char *curr = p; curr < p + SIZE; curr += PAGE_SIZE)
 			*(unsigned long *)curr = (unsigned long) curr;
 	} else {
 		memset(p, 0xff, SIZE);
 	}
 	printf("pause\n");
 	char x; read(0, &x, 1);
 	for (char *curr = p; curr < p + SIZE; curr += PAGE_SIZE)
 		*(unsigned long *)curr = (unsigned long) curr;
 	return 0;
 }
====

Before this commit this the wrprotect fault latency on matched
pages (2 runs):

[4K, 8K)               1 |@@@@                                                |
[8K, 16K)             12 |@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@|
[16K, 32K)             1 |@@@@                                                |
[32K, 64K)             1 |@@@@                                                |
[64K, 128K)            0 |                                                    |
[128K, 256K)           0 |                                                    |
[256K, 512K)           1 |@@@@                                                |

[4K, 8K)               1 |@@@@                                                |
[8K, 16K)             12 |@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@|
[16K, 32K)             2 |@@@@@@@@                                            |
[32K, 64K)             0 |                                                    |
[64K, 128K)            0 |                                                    |
[128K, 256K)           1 |@@@@                                                |

and on unmatched pages (2 runs):

[1K, 2K)              13 |@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@|
[2K, 4K)               1 |@@@@                                                |
[4K, 8K)               0 |                                                    |
[8K, 16K)              0 |                                                    |
[16K, 32K)             1 |@@@@                                                |
[32K, 64K)             0 |                                                    |
[64K, 128K)            1 |@@@@                                                |

[1K, 2K)              14 |@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@|
[2K, 4K)               1 |@@@                                                 |
[4K, 8K)               0 |                                                    |
[8K, 16K)              0 |                                                    |
[16K, 32K)             0 |                                                    |
[32K, 64K)             0 |                                                    |
[64K, 128K)            0 |                                                    |
[128K, 256K)           1 |@@@                                                 |

After this commit this the wrprotect fault latency on matched pages (2 runs):

[8K, 16K)             11 |@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@|
[16K, 32K)             4 |@@@@@@@@@@@@@@@@@@                                  |
[32K, 64K)             0 |                                                    |
[64K, 128K)            0 |                                                    |
[128K, 256K)           0 |                                                    |
[256K, 512K)           0 |                                                    |
[512K, 1M)             1 |@@@@                                                |

[4K, 8K)               1 |@@@@                                                |
[8K, 16K)             13 |@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@|
[16K, 32K)             1 |@@@@                                                |
[32K, 64K)             0 |                                                    |
[64K, 128K)            0 |                                                    |
[128K, 256K)           1 |@@@@                                                |

and on unmatched pages (2 runs):

[8K, 16K)             13 |@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@|
[16K, 32K)             1 |@@@@                                                |
[32K, 64K)             1 |@@@@                                                |
[64K, 128K)            0 |                                                    |
[128K, 256K)           1 |@@@@                                                |

[8K, 16K)             12 |@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@|
[16K, 32K)             3 |@@@@@@@@@@@@@                                       |
[32K, 64K)             0 |                                                    |
[64K, 128K)            0 |                                                    |
[128K, 256K)           1 |@@@@                                                |

This commit makes the wrprotect faults on matched and unmatched KSM
pages indistinguishable (8-16usec average latency on cold caches,
without including kernel entry/exit).

Before this commit COWs on unmatched patches would run much faster
with a lower 1-2usec latency, so they would still have been easy to
detect and there would have been no much point in the
ksm_unmatched_random_distribution wrprotection effort in the first
place.

NOTE: because unmatched pages (wrprotected by KSM with
ksm_unmatched_random_distribution > 0) aren't converted to PageKsm and
they remain PageAnon, the COW fault is forbidden to alter the virt to
phys translation or it'd break the coherency of the GUP pins. For
example a FOLL_LONGTERM with FOLL_MM_SYNC (or the obsoleted
FOLL_WRITE|FOLL_FORCE that predated FOLL_MM_SYNC) would have
unconditionally unshared a PageKsm before taking the long term
FOLL_PIN, but it won't do that for a PageAnon && !PageKsm.

The PageKsm is specially costrued so that it is always safe to break
COW without losing synchronicity regardless of any outstanding GUP
pin. In fact a PageKsm is created only if there are no outstanding GUP
pins (the page_count vs mapcount check happens with the PT lock held
and also while the pgtable is non present to keep gup/pin-fast at
bay).

The previous paragraph explains why the COW-break must be simulated
without actually replacing the page to avoid reintroducing the defects
of 09854ba94c6aad7886996bfbee2530b3d8a7f4f4 once again to the PageAnon
&& !PageKsm COWs happening in VM_MERGEABLE vmas.

It would be possible to mark the unmatched pages or to disable the
FAULT_FLAG_FAKECOW behavior completely when
ksm_unmatched_random_distribution is zero to avoid the "slower COW"
behavior when not strictly necessary. It's not urgent though because
the most common use cases of KSM are unlikely to benefit from such an
optimization. KSM is primarily used when fork() has been ruled out as
an option to share Anonymous memory securely across different
processes as in zygote, and KDE Plasma (i.e. KProcess instead of
fork+exec), without paying for extra background runtime CPU costs and
extra RAM costs for non linear rmaps (rmap_items).

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>