KSM: if the page isn't mapped page_anon_vma(page) can be stale
Only reuse the swapcache for the next KSM swapin if the swapcache is
currently mapped, otherwise the non linearity of the
rmap_item->anon_vmas of KSM pages implies the previous KSM swapin that
allocated the Anon swapcache, could not share the same anon_vma root
with the new KSM swapin. In turn the anon_vma could have been freed,
if the swapcache is not mapped anymore at the time of the fault.

This should fix the crash reported here:
https://lkml.kernel.org/r/81416741-d11c-61d4-26f9-4546dd874581@huawei.com

Note: commit 09854ba94c6aad7886996bfbee2530b3d8a7f4f4 in this case has
nothing to do with it, this is a longstanding issue that always
existed in KSM code.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>