mm: gup: enable PageAnonGup to reduce the false positive copies in fork
This reduces the false positives COWs during fork to the intersections
of the anon pages that have ever been pinned with the ones that still
have the page count elevated for some reason and that aren't currently
being shared.

The original implementation of the gup page bitflag was proposed in
the link. This however handles hugetlbfs and THP too so a simple page
bitflag isn't enough and a compound one is required too.

hugetlbfs only needs to reflect the "anon gup" bit in the compound
field, the subpage bitflag is never set.

THP has to reflect "anon gup" bit in the compound field as well as the
subpage bit.

To monitor the effect:

   bpftrace -e 'kretprobe:page_needs_cow_for_dma { @tot[retval] = count() } kretprobe:page_needs_cow_for_dma { @tot[retval] = count() }'

Running the above during sddm plasma login and then starting firefox
on Fedora the output is:

[..]
@tot[1]: 80
@tot[0]: 1037304

The false positives COWs are 80 over a total of >1M COWs. THP in the
test was set to "always" in the kernel config.

Even if all other things would be equal, the extra false positive
copies, if they might ever happen, are preferable to deal with in
fork() than the COW fault, because:

1) there will be at most one extra false positive copy per page no
   matter the amount of later mprotect and other mm activity.

2) if the false positive copy fails the allocation we can fail
   gracefully with -ENOMEM in fork() unlike in the COW fault which is
   forced to segfault.

3) fork() is the only place where we can count the mappings and skip
   the extra copies for virtually splitted THP. The COW fault doesn't
   hold the right lock for that.

Link: lkml.kernel.org/r/20090311165833.GI27823@random.random
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>