mm: gup: enable PageAnonGup to reduce the false positive copies in fork
This reduces the false positives COWs during fork to the intersections
of the anon pages that have ever been pinned with the ones that still
have the page count elevated for some reason and that aren't currently
being shared.

The original implementation of the gup page bitflag was proposed in
the link. This however handles hugetlbfs and THP too so a simple page
bitflag isn't enough and a compound one is required too.

hugetlbfs only needs to reflect the "anon gup" bit in the compound
field, the subpage bitflag is never set.

THP has to reflect "anon gup" bit in the compound field as well as the
subpage bit.

To monitor the effect:

   bpftrace -e 'kretprobe:page_needs_cow_for_dma /retval/ { @[comm] = count() }'

Link: lkml.kernel.org/r/20090311165833.GI27823@random.random
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>